+++ server/Session.js
@@ -1,8 +1,9 @@
 class Session {
-  constructor(id, chatbotState, alertType, createdAt, updatedAt, incidentCategory, respondedAt, respondedByPhoneNumber, location) {
+  constructor(id, chatbotState, alertType, numberOfAlerts, createdAt, updatedAt, incidentCategory, respondedAt, respondedByPhoneNumber, location) {
     this.id = id
     this.chatbotState = chatbotState
     this.alertType = alertType
+    this.numberOfAlerts = numberOfAlerts
     this.createdAt = createdAt
     this.updatedAt = updatedAt
     this.incidentCategory = incidentCategory