+++ server/test/integration/BraveAlerterConfiguratorTest/getReturnMessageToOtherResponderPhoneNumbersTest.js
@@ -6,7 +6,7 @@ const { before, describe, it } = require('mocha')
 const { CHATBOT_STATE } = require('brave-alert-lib')
 const BraveAlerterConfigurator = require('../../../BraveAlerterConfigurator')
 
-describe('BraveAlerterConfigurator.js unit tests: getReturnMessageToOtherResponderPhoneNumbers', () => {
+describe('BraveAlerterConfigurator.js integration tests: getReturnMessageToOtherResponderPhoneNumbers', () => {
   before(() => {
     const braveAlerterConfigurator = new BraveAlerterConfigurator()
     const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
@@ -33,6 +33,14 @@ describe('BraveAlerterConfigurator.js unit tests: getReturnMessageToOtherRespond
     expect(returnMessage).to.equal(`Another Responder has acknowledged this request. (You don't need to respond to this message.)`)
   })
 
+  it('should get message when STARTED => RESET', () => {
+    const returnMessage = this.alertStateMachine.getReturnMessageToOtherResponderPhoneNumbers('en', CHATBOT_STATE.STARTED, CHATBOT_STATE.RESET)
+
+    expect(returnMessage).to.equal(
+      "Another Responder has reset the Brave Sensor.\n\nThe sensor will return to normal operation after about ten minutes have passed. If you have not checked on this location yet, please do so.\n\nThis session is now complete. (You don't need to respond to this message.)",
+    )
+  })
+
   it('should get message when WAITING_FOR_REPLY => WAITING_FOR_CATEGORY', () => {
     const returnMessage = this.alertStateMachine.getReturnMessageToOtherResponderPhoneNumbers(
       'en',
@@ -42,6 +50,18 @@ describe('BraveAlerterConfigurator.js unit tests: getReturnMessageToOtherRespond
     )
 
     expect(returnMessage).to.equal(`Another Responder has acknowledged this request. (You don't need to respond to this message.)`)
+  })
+
+  it('should get message when WAITING_FOR_REPLY => RESET', () => {
+    const returnMessage = this.alertStateMachine.getReturnMessageToOtherResponderPhoneNumbers(
+      'en',
+      CHATBOT_STATE.WAITING_FOR_REPLY,
+      CHATBOT_STATE.RESET,
+    )
+
+    expect(returnMessage).to.equal(
+      "Another Responder has reset the Brave Sensor.\n\nThe sensor will return to normal operation after about ten minutes have passed. If you have not checked on this location yet, please do so.\n\nThis session is now complete. (You don't need to respond to this message.)",
+    )
   })
 
   it('should get no message when WAITING_FOR_CATEGORY => WAITING_FOR_CATEGORY', () => {
@@ -74,6 +94,12 @@ describe('BraveAlerterConfigurator.js unit tests: getReturnMessageToOtherRespond
       CHATBOT_STATE.COMPLETED,
       ['Cat0'],
     )
+
+    expect(returnMessage).to.equal(null)
+  })
+
+  it('should get no message when RESET => RESET', () => {
+    const returnMessage = this.alertStateMachine.getReturnMessageToOtherResponderPhoneNumbers('en', CHATBOT_STATE.RESET, CHATBOT_STATE.RESET)
 
     expect(returnMessage).to.equal(null)
   })