+++ server/testingHelpers.js
@@ -96,6 +96,7 @@ function sessionFactory(overrides = {}) {
     overrides.id !== undefined ? overrides.id : 'd91593b4-25ce-11ec-9621-0242ac130002',
     overrides.chatbotState !== undefined ? overrides.chatbotState : CHATBOT_STATE.COMPLETED,
     overrides.alertType !== undefined ? overrides.alertType : ALERT_TYPE.SENSOR_STILLNESS,
+    overrides.numberOfAlerts !== undefined ? overrides.numberOfAlerts : 1,
     overrides.createdAt !== undefined ? overrides.createdAt : new Date('2021-10-05T20:20:20.000Z'),
     overrides.updatedAt !== undefined ? overrides.updatedAt : new Date('2021-10-05T20:20:55.000Z'),
     overrides.incidentCategory !== undefined ? overrides.incidentCategory : 'Overdose',
@@ -179,6 +180,16 @@ function mockResponse(sandbox) {
   return res
 }
 
+function mockBraveAlerter(sandbox) {
+  const braveAlerter = {}
+
+  braveAlerter.sendSingleAlert = sandbox.stub()
+  braveAlerter.sendAlertSessionUpdate = sandbox.stub()
+  braveAlerter.startAlertSession = sandbox.stub()
+
+  return braveAlerter
+}
+
 module.exports = {
   firmwareAlert,
   getRandomArbitrary,
@@ -192,4 +203,5 @@ module.exports = {
   sessionDBFactory,
   sessionFactory,
   mockResponse,
+  mockBraveAlerter,
 }